+++ docs/metrics/extending-the-sdk/MyExporter.cs
@@ -43,10 +43,9 @@ internal class MyExporter : BaseExporter<Metric>
             }
 
             sb.Append($"{record}");
-            sb.Append(')');
         }
 
-        Console.WriteLine($"{this.name}.Export([{sb.ToString()}])");
+        Console.WriteLine($"{this.name}.Export([{sb}])");
         return ExportResult.Success;
     }