+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpExporterOptions.cs
@@ -31,6 +31,10 @@ namespace OpenTelemetry.Exporter
         internal const string EndpointEnvVarName = "OTEL_EXPORTER_OTLP_ENDPOINT";
         internal const string HeadersEnvVarName = "OTEL_EXPORTER_OTLP_HEADERS";
         internal const string TimeoutEnvVarName = "OTEL_EXPORTER_OTLP_TIMEOUT";
+        internal const string ProtocolEnvVarName = "OTEL_EXPORTER_OTLP_PROTOCOL";
+
+        internal const string TracesExportPath = "v1/traces";
+        internal const string MetricsExportPath = "v1/metrics";
 
         /// <summary>
         /// Initializes a new instance of the <see cref="OtlpExporterOptions"/> class.
@@ -70,6 +74,20 @@ namespace OpenTelemetry.Exporter
                         OpenTelemetryProtocolExporterEventSource.Log.FailedToParseEnvironmentVariable(TimeoutEnvVarName, timeoutEnvVar);
                     }
                 }
+
+                string protocolEnvVar = Environment.GetEnvironmentVariable(ProtocolEnvVarName);
+                if (!string.IsNullOrEmpty(protocolEnvVar))
+                {
+                    var protocol = protocolEnvVar.ToOtlpExportProtocol();
+                    if (protocol.HasValue)
+                    {
+                        this.Protocol = protocol.Value;
+                    }
+                    else
+                    {
+                        OpenTelemetryProtocolExporterEventSource.Log.UnsupportedProtocol(protocolEnvVar);
+                    }
+                }
             }
             catch (SecurityException ex)
             {
@@ -80,10 +98,9 @@ namespace OpenTelemetry.Exporter
         }
 
         /// <summary>
-        /// Gets or sets the target to which the exporter is going to send traces.
-        /// Must be a valid Uri with scheme (http) and host, and
-        /// may contain a port and path. Secure connection(https) is not
-        /// supported.
+        /// Gets or sets the target to which the exporter is going to send telemetry.
+        /// Must be a valid Uri with scheme (http or https) and host, and
+        /// may contain a port and path. The default value is http://localhost:4317.
         /// </summary>
         public Uri Endpoint { get; set; } = new Uri("http://localhost:4317");
 
@@ -94,9 +111,14 @@ namespace OpenTelemetry.Exporter
         public string Headers { get; set; }
 
         /// <summary>
-        /// Gets or sets the max waiting time (in milliseconds) for the backend to process each span batch. The default value is 10000.
+        /// Gets or sets the max waiting time (in milliseconds) for the backend to process each batch. The default value is 10000.
         /// </summary>
         public int TimeoutMilliseconds { get; set; } = 10000;
+
+        /// <summary>
+        /// Gets or sets the the OTLP transport protocol. Supported values: Grpc and HttpProtobuf.
+        /// </summary>
+        public OtlpExportProtocol Protocol { get; set; } = OtlpExportProtocol.Grpc;
 
         /// <summary>
         /// Gets or sets the export processor type to be used with the OpenTelemetry Protocol Exporter. The default value is <see cref="ExportProcessorType.Batch"/>.