+++ src/OpenTelemetry/Metrics/BaseExportingMetricReader.cs
@@ -15,6 +15,8 @@
 // </copyright>
 
 using System;
+using System.Diagnostics;
+using System.Threading;
 
 namespace OpenTelemetry.Metrics
 {
@@ -43,7 +45,27 @@ namespace OpenTelemetry.Metrics
                 var attr = (ExportModesAttribute)attributes[attributes.Length - 1];
                 this.supportedExportModes = attr.Supported;
             }
+
+            if (exporter is IPullMetricExporter pullExporter)
+            {
+                if (this.supportedExportModes.HasFlag(ExportModes.Push))
+                {
+                    pullExporter.Collect = this.Collect;
+                }
+                else
+                {
+                    pullExporter.Collect = (timeoutMilliseconds) =>
+                    {
+                        using (PullMetricScope.Begin())
+                        {
+                            return this.Collect(timeoutMilliseconds);
+                        }
+                    };
+                }
+            }
         }
+
+        internal BaseExporter<Metric> Exporter => this.exporter;
 
         protected ExportModes SupportedExportModes => this.supportedExportModes;
 
@@ -56,13 +78,46 @@ namespace OpenTelemetry.Metrics
         /// <inheritdoc/>
         protected override bool ProcessMetrics(Batch<Metric> metrics, int timeoutMilliseconds)
         {
+            // TODO: Do we need to consider timeout here?
             return this.exporter.Export(metrics) == ExportResult.Success;
+        }
+
+        /// <inheritdoc />
+        protected override bool OnCollect(int timeoutMilliseconds)
+        {
+            if (this.supportedExportModes.HasFlag(ExportModes.Push))
+            {
+                return base.OnCollect(timeoutMilliseconds);
+            }
+
+            if (this.supportedExportModes.HasFlag(ExportModes.Pull) && PullMetricScope.IsPullAllowed)
+            {
+                return base.OnCollect(timeoutMilliseconds);
+            }
+
+            // TODO: add some error log
+            return false;
         }
 
         /// <inheritdoc />
         protected override bool OnShutdown(int timeoutMilliseconds)
         {
-            return this.exporter.Shutdown(timeoutMilliseconds);
+            var result = true;
+
+            if (timeoutMilliseconds == Timeout.Infinite)
+            {
+                result = this.Collect(Timeout.Infinite) && result;
+                result = this.exporter.Shutdown(Timeout.Infinite) && result;
+            }
+            else
+            {
+                var sw = Stopwatch.StartNew();
+                result = this.Collect(timeoutMilliseconds) && result;
+                var timeout = timeoutMilliseconds - sw.ElapsedMilliseconds;
+                result = this.exporter.Shutdown((int)Math.Max(timeout, 0)) && result;
+            }
+
+            return result;
         }
 
         /// <inheritdoc/>
@@ -77,6 +132,11 @@ namespace OpenTelemetry.Metrics
             {
                 try
                 {
+                    if (this.exporter is IPullMetricExporter pullExporter)
+                    {
+                        pullExporter.Collect = null;
+                    }
+
                     this.exporter.Dispose();
                 }
                 catch (Exception)