+++ src/OpenTelemetry.Exporter.Prometheus/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## 1.2.0-beta1
+
+Released 2021-Oct-08
+
+## 1.2.0-alpha4
+
+Released 2021-Sep-23
+
 ## 1.2.0-alpha3
 
 Released 2021-Sep-13