+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+## 1.2.0-beta1
+
+Released 2021-Oct-08
+
+* Added .NET 5.0 target and threading optimizations
+  ([#2405](https://github.com/open-telemetry/opentelemetry-dotnet/issues/2405))
+
+## 1.2.0-alpha4
+
+Released 2021-Sep-23
+
 ## 1.2.0-alpha3
 
 Released 2021-Sep-13